Modified parameters for numbers of passes
Now creates all drops for all passes. This way we don't create a series
of passes with a series of drops, but rather only a series of drops.

The number of drop is * by number of passes / 2 because we want to
simulate a decreasing number of drop points per pass. This equation will
give the same number as the original Ruby algo I am porting.